+++ gems/aws-sdk-appconfig/lib/aws-sdk-appconfig/types.rb
@@ -304,6 +304,20 @@ module Aws::AppConfig
     #   `AWS.Freeform`
     #   @return [String]
     #
+    # @!attribute [rw] kms_key_arn
+    #   The Amazon Resource Name of the Key Management Service key to
+    #   encrypt new configuration data versions in the AppConfig hosted
+    #   configuration store. This attribute is only used for `hosted`
+    #   configuration types. To encrypt data managed in other configuration
+    #   stores, see the documentation for how to specify an KMS key for that
+    #   particular service.
+    #   @return [String]
+    #
+    # @!attribute [rw] kms_key_identifier
+    #   The Key Management Service key identifier (key ID, key alias, or key
+    #   ARN) provided when the resource was created or updated.
+    #   @return [String]
+    #
     # @see http://docs.aws.amazon.com/goto/WebAPI/appconfig-2019-10-09/ConfigurationProfile AWS API Documentation
     #
     class ConfigurationProfile < Struct.new(
@@ -314,7 +328,9 @@ module Aws::AppConfig
       :location_uri,
       :retrieval_role_arn,
       :validators,
-      :type)
+      :type,
+      :kms_key_arn,
+      :kms_key_identifier)
       SENSITIVE = []
       include Aws::Structure
     end
@@ -492,6 +508,16 @@ module Aws::AppConfig
     #   `AWS.Freeform`
     #   @return [String]
     #
+    # @!attribute [rw] kms_key_identifier
+    #   The identifier for an Key Management Service key to encrypt new
+    #   configuration data versions in the AppConfig hosted configuration
+    #   store. This attribute is only used for `hosted` configuration types.
+    #   The identifier can be an KMS key ID, alias, or the Amazon Resource
+    #   Name (ARN) of the key ID or alias. To encrypt data managed in other
+    #   configuration stores, see the documentation for how to specify an
+    #   KMS key for that particular service.
+    #   @return [String]
+    #
     # @see http://docs.aws.amazon.com/goto/WebAPI/appconfig-2019-10-09/CreateConfigurationProfileRequest AWS API Documentation
     #
     class CreateConfigurationProfileRequest < Struct.new(
@@ -502,7 +528,8 @@ module Aws::AppConfig
       :retrieval_role_arn,
       :validators,
       :tags,
-      :type)
+      :type,
+      :kms_key_identifier)
       SENSITIVE = []
       include Aws::Structure
     end
@@ -971,9 +998,8 @@ module Aws::AppConfig
     #   @return [String]
     #
     # @!attribute [rw] kms_key_identifier
-    #   The KMS key identifier (key ID, key alias, or key ARN). AppConfig
-    #   uses this ID to encrypt the configuration data using a customer
-    #   managed key.
+    #   The Key Management Service key identifier (key ID, key alias, or key
+    #   ARN) provided when the resource was created or updated.
     #   @return [String]
     #
     # @!attribute [rw] version_label
@@ -1695,6 +1721,12 @@ module Aws::AppConfig
     #   A user-defined label for an AppConfig hosted configuration version.
     #   @return [String]
     #
+    # @!attribute [rw] kms_key_arn
+    #   The Amazon Resource Name of the Key Management Service key that was
+    #   used to encrypt this specific version of the configuration data in
+    #   the AppConfig hosted configuration store.
+    #   @return [String]
+    #
     # @see http://docs.aws.amazon.com/goto/WebAPI/appconfig-2019-10-09/HostedConfigurationVersion AWS API Documentation
     #
     class HostedConfigurationVersion < Struct.new(
@@ -1704,7 +1736,8 @@ module Aws::AppConfig
       :description,
       :content,
       :content_type,
-      :version_label)
+      :version_label,
+      :kms_key_arn)
       SENSITIVE = [:content]
       include Aws::Structure
     end
@@ -1740,6 +1773,12 @@ module Aws::AppConfig
     #   A user-defined label for an AppConfig hosted configuration version.
     #   @return [String]
     #
+    # @!attribute [rw] kms_key_arn
+    #   The Amazon Resource Name of the Key Management Service key that was
+    #   used to encrypt this specific version of the configuration data in
+    #   the AppConfig hosted configuration store.
+    #   @return [String]
+    #
     # @see http://docs.aws.amazon.com/goto/WebAPI/appconfig-2019-10-09/HostedConfigurationVersionSummary AWS API Documentation
     #
     class HostedConfigurationVersionSummary < Struct.new(
@@ -1748,7 +1787,8 @@ module Aws::AppConfig
       :version_number,
       :description,
       :content_type,
-      :version_label)
+      :version_label,
+      :kms_key_arn)
       SENSITIVE = []
       include Aws::Structure
     end
@@ -2361,6 +2401,16 @@ module Aws::AppConfig
     #   A list of methods for validating the configuration.
     #   @return [Array<Types::Validator>]
     #
+    # @!attribute [rw] kms_key_identifier
+    #   The identifier for a Key Management Service key to encrypt new
+    #   configuration data versions in the AppConfig hosted configuration
+    #   store. This attribute is only used for `hosted` configuration types.
+    #   The identifier can be an KMS key ID, alias, or the Amazon Resource
+    #   Name (ARN) of the key ID or alias. To encrypt data managed in other
+    #   configuration stores, see the documentation for how to specify an
+    #   KMS key for that particular service.
+    #   @return [String]
+    #
     # @see http://docs.aws.amazon.com/goto/WebAPI/appconfig-2019-10-09/UpdateConfigurationProfileRequest AWS API Documentation
     #
     class UpdateConfigurationProfileRequest < Struct.new(
@@ -2369,7 +2419,8 @@ module Aws::AppConfig
       :name,
       :description,
       :retrieval_role_arn,
-      :validators)
+      :validators,
+      :kms_key_identifier)
       SENSITIVE = []
       include Aws::Structure
     end