+++ gems/aws-sdk-mediatailor/lib/aws-sdk-mediatailor.rb
@@ -52,6 +52,6 @@ require_relative 'aws-sdk-mediatailor/customizations'
 # @!group service
 module Aws::MediaTailor
 
-  GEM_VERSION = '1.70.0'
+  GEM_VERSION = '1.71.0'
 
 end