+++ gems/aws-partitions/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.833.0 (2023-10-04)
+------------------
+
+* Feature - Added support for enumerating regions for `Aws::DataZone`.
+
 1.832.0 (2023-10-03)
 ------------------