+++ gems/aws-sdk-mediatailor/CHANGELOG.md
@@ -1,6 +1,11 @@
 Unreleased Changes
 ------------------
 
+1.71.0 (2023-10-04)
+------------------
+
+* Feature - Updates DescribeVodSource to include a list of ad break opportunities in the response
+
 1.70.0 (2023-09-27)
 ------------------